backend/requirements.txt
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
numpy==1.24.3
psycopg2-binary==2.9.6
huggingface_hub==0.15.1
